Use bash in utils/latex.sh and the "-e" option to echo
Because on Debian /bin/sh is dash and the echo builtin in dash
correctly prints "\appendix" when executing "echo '\\appendix' but
when /bin/sh is bash like it is on Fedora, then this doesn't work and
"echo '\\appendix'" prints "\\appendix" if the "-e" option is not
used. Using the "-e" option to echo and using bash always should make
it work on all systems.